+++ src/servers/app/DefaultDecorator.cpp
@@ -350,7 +350,7 @@ DefaultDecorator::_DoLayout()
 
 	bool hasTab = false;
 
-	switch (Look()) {
+	switch ((int)Look()) {
 		case B_MODAL_WINDOW_LOOK:
 			fBorderWidth = 5;
 			break;
@@ -511,7 +511,7 @@ DefaultDecorator::_DrawFrame(BRect invalid)
 
 	// Draw the border frame
 	BRect r = BRect(fTopBorder.LeftTop(), fBottomBorder.RightBottom());
-	switch (fLook) {
+	switch ((int)fLook) {
 		case B_TITLED_WINDOW_LOOK:
 		case B_DOCUMENT_WINDOW_LOOK:
 		case B_MODAL_WINDOW_LOOK:
@@ -655,7 +655,7 @@ DefaultDecorator::_DrawFrame(BRect invalid)
 		ComponentColors colors;
 		_GetComponentColors(COMPONENT_RESIZE_CORNER, colors);
 
-		switch (fLook) {
+		switch ((int)fLook) {
 			case B_DOCUMENT_WINDOW_LOOK:
 			{
 				if (!invalid.Intersects(r))
@@ -989,7 +989,7 @@ DefaultDecorator::_ResizeBy(BPoint offset, BRegion* dirty)
 	// Handle invalidation of resize rect
 	if (dirty && !(fFlags & B_NOT_RESIZABLE)) {
 		BRect realResizeRect;
-		switch (fLook) {
+		switch ((int)fLook) {
 			case B_DOCUMENT_WINDOW_LOOK:
 				realResizeRect = fResizeRect;
 				// Resize rect at old location
